Chore: Move collection queries to inside the collection component
