+++ react/components/CollectionsTranslation/CollectionTranslation.tsx
@@ -13,10 +13,10 @@ import { FormattedMessage, useIntl } from 'react-intl'
 import useCatalogQuery from '../../hooks/useCatalogQuery'
 import { useLocaleSelector } from '../LocaleSelector'
 import ErrorHandler from '../ErrorHandler'
-import getCollectionById from '../../graphql/getCollections.gql'
-import getAllCollections from '../../graphql/getAllCollections.gql'
+import getCollectionById from './graphql/getCollections.gql'
+import getAllCollections from './graphql/getAllCollections.gql'
 import CollectionsForm from './CollectionForm'
-import QUERY_MESSAGES from '../../graphql/messages.gql'
+import QUERY_MESSAGES from './graphql/messages.gql'
 import { formatCollectionFromMessages, parseJSONToXLS } from '../../utils'
 import CollectionImportModal from './CollectionImportModal'
