+++ react/components/CollectionsTranslation/graphql/getAllCollections.gql
