+++ react/components/CollectionsTranslation/graphql/getCollections.gql
