+++ react/components/CollectionsTranslation/graphql/messages.gql
