+++ react/components/CollectionsTranslation/graphql/translateMessages.gql
