+++ react/components/CollectionsTranslation/graphql/uploadCollectionTranslation.gql
